+++ src/creators/character.js
@@ -1,13 +1,13 @@
-// // @flow
+// @flow
 
-// import type { EmbedProps } from '../bootstrap';
+import type { EmbedProps } from '../bootstrap';
 
-// import React from 'react';
-// import Character from '../components/Character';
+import React from 'react';
+import { CharacterPreview } from 'armory-component-ui';
 
-// import { makeAttribute } from '../bootstrap';
+import { makeAttribute } from '../bootstrap';
 
-// export default function (element: HTMLElement) {
-//   const name = element.getAttribute(makeAttribute('name')) || '';
-//   return (props: EmbedProps) => <Character {...props} name={name} />;
-// }
+export default function (element: HTMLElement) {
+  const name = element.getAttribute(makeAttribute('name')) || '';
+  return (props: EmbedProps) => <CharacterPreview {...props} name={name} />;
+}