+++ src/creators/items.js
@@ -12,14 +12,28 @@ const extractAttr = (element, key, ids) => {
       return {};
     }
 
-    return { [id]: parseInt(rawId, 10) };
+    return {
+      [id]: rawId.includes(',')
+        ? rawId.split(',').map((num) => parseInt(num, 10))
+        : parseInt(rawId, 10),
+    };
   })
   .reduce((map, next) => Object.assign(map, next), {});
 };
 
 export default function (element: HTMLElement, ids: Array<number>) {
-  const statIds = extractAttr(element, 'stat', ids);
-  const skinIds = extractAttr(element, 'skin', ids);
+  const stats = extractAttr(element, 'stat', ids);
+  const skins = extractAttr(element, 'skin', ids);
+  const upgrades = extractAttr(element, 'upgrades', ids);
+  const infusions = extractAttr(element, 'infusions', ids);
 
-  return (props: EmbedProps) => <Items {...props} ids={ids} statIds={statIds} skinIds={skinIds} />;
+  return (props: EmbedProps) =>
+    <Items
+      {...props}
+      upgradeIds={upgrades}
+      infusionIds={infusions}
+      ids={ids}
+      statIds={stats}
+      skinIds={skins}
+    />;
 }