+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Added
+- Character embed
+- Upgrades to item embed
+- Infusions to item embed
 
 ## [0.2.5][] - 2017-10-19
 ### Fixed